printk: Add ignore_per_console_loglevel module parameter
With per-console loglevels now functional, there are scenarios where
administrators need to quickly disable them and revert to the global
console_loglevel for all consoles. This is useful for debugging message
emission issues or when needing to break glass in emergencies.

This commit introduces the ignore_per_console_loglevel parameter which
can be set via the kernel command line or at runtime through
/sys/module/printk/parameters/ignore_per_console_loglevel. When set,
per-console loglevels are bypassed and the global console_loglevel
is used for all consoles.

The variable is defined outside CONFIG_PRINTK so that the sysrq
infrastructure (which will use it in a subsequent commit) can access
it unconditionally.

The console_effective_loglevel_source() function is updated to check
this flag, returning LLS_GLOBAL instead of LLS_LOCAL when the flag
is set.

Reviewed-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Chris Down <chris@chrisdown.name>